requirements.txt
streamlit>=1.28.0
groq>=0.4.1
python-dotenv>=1.0.0
phonenumbers>=8.13.0
pandas>=2.0.0
hashlib2>=1.0.1
regex>=2023.10.3